requirements.txt
numpy==2.2.4
matplotlib==3.10.1
pillow==11.1.0
